02_movies_admin/requirements.txt
django==3.2
flake8==4.0
django-admin==2.0
python-dotenv==0.20
psycopg2==2.9.3
django-split-settings==1.1.0